+++ package.json
@@ -19,6 +19,7 @@
     "react-bootstrap-icons": "^1.7.2",
     "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
     "styled-components": "^5.3.3",
     "typescript": "^4.6.2",
