+++ src/components/Header.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./header.css";
+import "../styles/header.css";
 
 const Header = () => {
   return (
