+++ src/pages/NotFound.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import "../styles/notfound.css";
+
+export default class NotFound extends React.Component {
+    render() {
+        return (
+            <div className="notfound">
+            <h1>Not Found</h1>
+            </div>
+        );
+    }
+}
