+++ src/styles/notfound.css
@@ -1,0 +1,3 @@
+h1 {
+    text-align: center;
+}
